+++ XPRising/Models/PlayerPreferences.cs
@@ -1,10 +1,16 @@
-﻿using XPRising.Systems;
+﻿using System.Runtime.Serialization;
+using System.Text.Json.Serialization;
+using XPRising.Systems;
 using XPRising.Transport;
 
 namespace XPRising.Models;
 
 public struct PlayerPreferences
 {
+    private const string DefaultXpColour = "#ffcc33";
+    private const string DefaultMasteryColour = "#ccff33";
+    private const string DefaultBloodMasteryColour = "#cc0000";
+
     public bool LoggingWanted = false;
     public bool LoggingExp = false;
     public bool LoggingMastery = false;
@@ -12,6 +18,12 @@ public struct PlayerPreferences
     public string Language = L10N.DefaultLanguage;
     public int TextSize = Plugin.DefaultTextSize;
     public Actions.BarState UIProgressDisplay = Actions.BarState.Active;
+    public string[] BarColours = [];
+    [JsonIgnore] public string XpBarColour => BarColours.ElementAtOrDefault(0) ?? DefaultXpColour;
+    [JsonIgnore] public string MasteryBarColour => BarColours.ElementAtOrDefault(1) ?? DefaultMasteryColour;
+    [JsonIgnore] public string BloodMasteryBarColour => BarColours.ElementAtOrDefault(2) ?? DefaultBloodMasteryColour;
+    [JsonIgnore]
+    public string[] BarColoursWithDefaults => new string[] {XpBarColour, MasteryBarColour, BloodMasteryBarColour};
 
     public PlayerPreferences()
     {